+++ src/base/QXmppConstants.cpp
@@ -133,6 +133,8 @@ const char* ns_mam = "urn:xmpp:mam:1";
 const char* ns_idle = "urn:xmpp:idle:1";
 // XEP-0333: Chat Markers
 const char* ns_chat_markers = "urn:xmpp:chat-markers:0";
+// XEP-0334: Message Processing Hints
+const char* ns_message_processing_hints = "urn:xmpp:hints";
 // XEP-0352: Client State Indication
 const char* ns_csi = "urn:xmpp:csi:0";
 // XEP-0363: HTTP File Upload